requirements.txt
streamlit>=1.31
python-dotenv>=1.0

openai-agents>=0.1.0

pydantic>=2.6